refactor: getAvailableApps() state handling
